+++ src/components/shared/SharedRundownTable.tsx
@@ -1307,9 +1307,9 @@ const SharedRundownTable = forwardRef<HTMLDivElement, SharedRundownTableProps>((
             <div 
               className="truncate overflow-hidden text-ellipsis whitespace-nowrap"
               style={{
-                width: `${parseInt(getColumnWidth(activeColumn)) - 16}px`,
-                minWidth: `${parseInt(getColumnWidth(activeColumn)) - 16}px`,
-                maxWidth: `${parseInt(getColumnWidth(activeColumn)) - 16}px`
+                width: `${parseFloat(getColumnWidth(activeColumn)) - 16}px`,
+                minWidth: `${parseFloat(getColumnWidth(activeColumn)) - 16}px`,
+                maxWidth: `${parseFloat(getColumnWidth(activeColumn)) - 16}px`
               }}
             >
               {activeColumn.name}
